pure ide: setup basic terminal functionalities
