+++ packages/legend-application-pure-ide/src/stores/EditorStore.ts
@@ -112,7 +112,8 @@ export class EditorStore implements CommandRegistrar {
   isMaxAuxPanelSizeSet = false;
   activeAuxPanelMode = AUX_PANEL_MODE.CONSOLE;
   readonly auxPanelDisplayState = new PanelDisplayState({
-    initial: 0,
+    // initial: 0,
+    initial: 300,
     default: 300,
     snap: 100,
   });
