+++ packages/legend-art/src/components/CJS__Icon.cts
@@ -26,6 +26,8 @@
  * See https://github.com/microsoft/TypeScript/issues/49298
  */
 export {
+  FaTrash as TrashIcon,
+  FaHistory as HistoryIcon,
   FaCrosshairs as CrosshairsIcon,
   FaLayerGroup as LayerGroupIcon,
   FaGhost as GhostIcon,
